scripts/dtc: Remove redundant YYLOC global declaration
See also: https://lkml.org/lkml/2020/4/1/1206

gcc 10 will default to -fno-common, which causes this error at link
time:

  (.text+0x0): multiple definition of `yylloc'; dtc-lexer.lex.o
    (symbol from plugin):(.text+0x0): first defined here

This is because both dtc-lexer as well as dtc-parser define the same
global symbol yyloc. Before with -fcommon those were merged into one
defintion. The proper solution would be to to mark this as "extern".

Signed-off-by: Dirk Mueller <dmueller@suse.com>
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
[robh: cherry-pick from upstream]
Cc: stable@vger.kernel.org
Signed-off-by: Rob Herring <robh@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Backport: mse1969@posteo.de
Whilst upstream and 4.4 indeed have removed the declaration completely,
in this kernel, it needs to be marked as "extern" to work.

Change-Id: I3b8c6ef698c42459d00a85de03207415860a8a9c