net: Drop native wireguard support
To use native wireguard support in 3.10, root access is necessary through
the wireguard app! For a security/privacy focused build, no features
requiring root access should be delivered.

The WG app also perfectly supports 'traditional' VPN tunnels, for which no
native kernel support is needed - the performance of this scenario is fine,
so dropping this support is no major loss.

Change-Id: I90752c5cbf64a6c1fe526a58d670c89bd831dcc0